rename path inference file
