+++ projects/optic/src/commands/capture/operations/path-inference.ts
